타입 수정

+++ src/main/java/com/dasom/whick/dto/PressureDto.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PressureDto {
-    private Integer pressure;  // 공기압 수치
+    private Float pressure;  // 공기압 수치
     private String status;     // 상태 (예: "정상", "위험")
 }
